requirement.txt
fastapi
uvicorn
sqlalchemy
pydantic-settings
psycopg2-binary
pwdlib[argon2]
pydantic[email]
PyJWT
